+++ next.config.js
@@ -1,27 +1,21 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: false,
+  output: "standalone",
+  eslint: { ignoreDuringBuilds: true },
+  typescript: { ignoreBuildErrors: true },
   experimental: {
     serverActions: true,
+    serverComponentsExternalPackages: ["@prisma/client", ".prisma/client"],
   },
-  output: "standalone",
-  staticPageGenerationTimeout: 180, // tingkatkan menjadi 3 menit
-  eslint: {
-    ignoreDuringBuilds: true,
+  webpack: (config, { isServer }) => {
+    if (isServer) {
+      config.externals = config.externals || [];
+      config.externals.push("@prisma/client");
+      config.externals.push(".prisma/client");
+    }
+    return config;
   },
-  // async headers() {
-  //   return [
-  //     {
-  //       source: "/(.*)",
-  //       headers: [
-  //         {
-  //           key: "Cache-Control",
-  //           value: "no-store, max-age=0",
-  //         },
-  //       ],
-  //     },
-  //   ];
-  // },
 };
 
 module.exports = nextConfig;