+++ src/app/api/event/tipe-acara/route.ts
@@ -33,7 +33,5 @@ export async function GET(request: Request) {
       },
       { status: 500 }
     );
-  } finally {
-    await prisma.$disconnect();
   }
 }