+++ src/app_modules/admin/donasi/fun/update/fun_status_publish.ts
@@ -33,12 +33,10 @@ export async function AdminDonasi_funUpdateStatusPublish(
   });
 
   if (!data) {
-    await prisma.$disconnect();
     return { status: 400, message: "Data tidak ditemukan" };
   }
 
   revalidatePath(RouterAdminDonasi.table_review);
-  await prisma.$disconnect();
 
   return {
     data: data,