+++ src/app_modules/donasi/fun/get/fun_check_status.tsx
@@ -9,9 +9,6 @@ export async function donasi_checkStatus({ id }: { id: string }) {
     },
   });
 
-  await prisma.$disconnect();
-
-
   if (checkStatus?.donasiMaster_StatusDonasiId == "2") return true;
   return false;